chore: Update test
