+++ extensions/warp-mp-ipfs/tests/friends.rs
@@ -1,38 +1,22 @@
+pub mod common;
 #[cfg(test)]
 mod test {
     use std::time::Duration;
 
+    use crate::common::{create_account, create_accounts};
     use futures::StreamExt;
-    use warp::crypto::DID;
-    use warp::multipass::identity::Identity;
-    use warp::multipass::{MultiPass, MultiPassEventKind};
-    use warp::tesseract::Tesseract;
-    use warp_mp_ipfs::config::Discovery;
-    use warp_mp_ipfs::ipfs_identity_temporary;
-
-    async fn create_account(
-        username: Option<&str>,
-        passphrase: Option<&str>,
-        context: Option<String>,
-    ) -> anyhow::Result<(Box<dyn MultiPass>, DID, Identity)> {
-        let tesseract = Tesseract::default();
-        tesseract.unlock(b"internal pass").unwrap();
-        let mut config = warp_mp_ipfs::config::MpIpfsConfig::development();
-        config.store_setting.discovery = Discovery::Provider(context);
-        config.store_setting.use_phonebook = false;
-        let mut account = ipfs_identity_temporary(Some(config), tesseract, None).await?;
-        let did = account.create_identity(username, passphrase).await?;
-        let identity = account.get_own_identity().await?;
-        Ok((Box::new(account), did, identity))
-    }
+    use warp::multipass::MultiPassEventKind;
 
     #[tokio::test]
     async fn add_friend() -> anyhow::Result<()> {
-        let (mut account_a, did_a, _) =
-            create_account(Some("JohnDoe"), None, Some("test::add_friend".into())).await?;
+        let accounts = create_accounts(vec![
+            (Some("JohnDoe"), None, Some("test::add_friend".into())),
+            (Some("JaneDoe"), None, Some("test::add_friend".into())),
+        ])
+        .await?;
 
-        let (mut account_b, did_b, _) =
-            create_account(Some("JaneDoe"), None, Some("test::add_friend".into())).await?;
+        let (mut account_a, did_a, _) = accounts.first().cloned().unwrap();
+        let (mut account_b, did_b, _) = accounts.last().cloned().unwrap();
 
         let mut subscribe_a = account_a.subscribe().await?;
         let mut subscribe_b = account_b.subscribe().await?;
@@ -66,11 +50,14 @@ mod test {
 
     #[tokio::test]
     async fn remove_friend() -> anyhow::Result<()> {
-        let (mut account_a, did_a, _) =
-            create_account(Some("JohnDoe"), None, Some("test::remove_friend".into())).await?;
+        let accounts = create_accounts(vec![
+            (Some("JohnDoe"), None, Some("test::remove_friend".into())),
+            (Some("JaneDoe"), None, Some("test::remove_friend".into())),
+        ])
+        .await?;
 
-        let (mut account_b, did_b, _) =
-            create_account(Some("JaneDoe"), None, Some("test::remove_friend".into())).await?;
+        let (mut account_a, did_a, _) = accounts.first().cloned().unwrap();
+        let (mut account_b, did_b, _) = accounts.last().cloned().unwrap();
 
         let mut subscribe_a = account_a.subscribe().await?;
         let mut subscribe_b = account_b.subscribe().await?;
@@ -126,11 +113,14 @@ mod test {
 
     #[tokio::test]
     async fn reject_friend() -> anyhow::Result<()> {
-        let (mut account_a, _, _) =
-            create_account(Some("JohnDoe"), None, Some("test::reject_friend".into())).await?;
+        let accounts = create_accounts(vec![
+            (Some("JohnDoe"), None, Some("test::reject_friend".into())),
+            (Some("JaneDoe"), None, Some("test::reject_friend".into())),
+        ])
+        .await?;
 
-        let (mut account_b, did_b, _) =
-            create_account(Some("JaneDoe"), None, Some("test::reject_friend".into())).await?;
+        let (mut account_a, _, _) = accounts.first().cloned().unwrap();
+        let (mut account_b, did_b, _) = accounts.last().cloned().unwrap();
 
         let mut subscribe_a = account_a.subscribe().await?;
         let mut subscribe_b = account_b.subscribe().await?;
@@ -164,11 +154,15 @@ mod test {
 
     #[tokio::test]
     async fn close_request() -> anyhow::Result<()> {
-        let (mut account_a, _, _) =
-            create_account(Some("JohnDoe"), None, Some("test::close_request".into())).await?;
+        let accounts = create_accounts(vec![
+            (Some("JohnDoe"), None, Some("test::close_request".into())),
+            (Some("JaneDoe"), None, Some("test::close_request".into())),
+        ])
+        .await?;
 
-        let (mut account_b, did_b, _) =
-            create_account(Some("JaneDoe"), None, Some("test::close_request".into())).await?;
+        let (mut account_a, _, _) = accounts.first().cloned().unwrap();
+        let (mut account_b, did_b, _) = accounts.last().cloned().unwrap();
+
         let mut subscribe_a = account_a.subscribe().await?;
         let mut subscribe_b = account_b.subscribe().await?;
 
@@ -213,11 +207,15 @@ mod test {
 
     #[tokio::test]
     async fn incoming_request() -> anyhow::Result<()> {
-        let (mut account_a, did_a, _) =
-            create_account(Some("JohnDoe"), None, Some("test::incoming_request".into())).await?;
+        let accounts = create_accounts(vec![
+            (Some("JohnDoe"), None, Some("test::incoming_request".into())),
+            (Some("JaneDoe"), None, Some("test::incoming_request".into())),
+        ])
+        .await?;
 
-        let (mut account_b, did_b, _) =
-            create_account(Some("JaneDoe"), None, Some("test::incoming_request".into())).await?;
+        let (mut account_a, did_a, _) = accounts.first().cloned().unwrap();
+        let (mut account_b, did_b, _) = accounts.last().cloned().unwrap();
+
         let mut subscribe_a = account_a.subscribe().await?;
         let mut subscribe_b = account_b.subscribe().await?;
 
@@ -252,11 +250,15 @@ mod test {
 
     #[tokio::test]
     async fn outgoing_request() -> anyhow::Result<()> {
-        let (mut account_a, _, _) =
-            create_account(Some("JohnDoe"), None, Some("test::outgoing_request".into())).await?;
+        let accounts = create_accounts(vec![
+            (Some("JohnDoe"), None, Some("test::outgoing_request".into())),
+            (Some("JaneDoe"), None, Some("test::outgoing_request".into())),
+        ])
+        .await?;
 
-        let (_account_b, did_b, _) =
-            create_account(Some("JaneDoe"), None, Some("test::outgoing_request".into())).await?;
+        let (mut account_a, _, _) = accounts.first().cloned().unwrap();
+        let (_, did_b, _) = accounts.last().cloned().unwrap();
+
         let mut subscribe_a = account_a.subscribe().await?;
 
         account_a.send_request(&did_b).await?;
@@ -279,11 +281,22 @@ mod test {
 
     #[tokio::test]
     async fn block_unblock_identity() -> anyhow::Result<()> {
-        let (mut account_a, did_a, _) =
-            create_account(None, None, Some("test::block_unblock_identity".into())).await?;
+        let accounts = create_accounts(vec![
+            (
+                Some("JohnDoe"),
+                None,
+                Some("test::block_unblock_identity".into()),
+            ),
+            (
+                Some("JaneDoe"),
+                None,
+                Some("test::block_unblock_identity".into()),
+            ),
+        ])
+        .await?;
 
-        let (mut account_b, did_b, _) =
-            create_account(None, None, Some("test::block_unblock_identity".into())).await?;
+        let (mut account_a, did_a, _) = accounts.first().cloned().unwrap();
+        let (mut account_b, did_b, _) = accounts.last().cloned().unwrap();
 
         let mut subscribe_a = account_a.subscribe().await?;
         let mut subscribe_b = account_b.subscribe().await?;
